fix: name structure
